requirements.txt
pygame-ce==2.3.1
customtkinter==5.2.2
moderngl==5.9.0